moved countdown to semgroup package
