+++ src/main/java/nl/tudelft/model/Game.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.util.Iterator;
 import java.util.LinkedList;
 
+import nl.tudelft.semgroup4.CountdownController;
 import nl.tudelft.semgroup4.GameEndedState;
 import nl.tudelft.semgroup4.Renderable;
 import nl.tudelft.semgroup4.ShopState;
